Commit conflict

+++ IGRP/src/nosi/webapps/igrp/pages/migrate/Migrate.java
@@ -14,7 +14,7 @@ public class Migrate extends Model{
 	@RParam(rParamName = "p_hostname")
 	private String hostname;
 	@RParam(rParamName = "p_port")
-	private float port;
+	private int port;
 	@RParam(rParamName = "p_nome_de_bade_dados")
 	private String nome_de_bade_dados;
 	@RParam(rParamName = "p_username")
@@ -45,10 +45,10 @@ public class Migrate extends Model{
 		return this.hostname;
 	}
 	
-	public void setPort(float port){
+	public void setPort(int port){
 		this.port = port;
 	}
-	public float getPort(){
+	public int getPort(){
 		return this.port;
 	}
